Move get_lifecycle_type to common
