+++ src/roadmap/common.py
@@ -11,6 +11,7 @@ from urllib.error import HTTPError
 from fastapi import HTTPException
 
 from roadmap.config import SETTINGS
+from roadmap.models import LifecycleType
 
 
 logger = logging.getLogger("uvicorn.error")
@@ -93,3 +94,35 @@ async def query_host_inventory(
         raise HTTPException(status_code=err.code, detail=err.msg)
 
     return data
+
+
+def get_lifecycle_type(products: list[dict[str, str]]) -> LifecycleType:
+    """Calculate lifecycle type based on the product ID.
+
+    https://downloads.corp.redhat.com/internal/products
+    https://github.com/RedHatInsights/rhsm-subscriptions/tree/main/swatch-product-configuration/src/main/resources/subscription_configs/RHEL
+
+    Mainline < EUS < E4S/EEUS < AUS
+
+    EUS --> 70, 73, 75
+    ELS --> 204
+    E4S/EEUS --> 241
+    AUS --> 251
+
+    """
+    ids = {item.get("id") for item in products}
+    type = LifecycleType.mainline
+
+    if any(id in ids for id in {"70", "73", "75"}):
+        type = LifecycleType.eus
+
+    if "204" in ids:
+        type = LifecycleType.els
+
+    if "241" in ids:
+        type = LifecycleType.e4s
+
+    if "251" in ids:
+        type = LifecycleType.aus
+
+    return type
